Solution to Clique Tree Calibrate
Previous commit mentioned a bug in computing messages. Turns out its not a bug, but a misunderstanding on my part. Calibration means neighboring cliques agree on the distribution of beliefs of the variables in the sepset, not the distribution of the messages. Normalized the messages at each step before passing then up/downstream. Computed final potentials.